Moved WP_Rocket\Traits\DeprecatedClass to WP_Rocket\deprecated\DeprecatedClassTrait
