+++ inc/deprecated/subscriber/Optimization/class-remove-query-string-subscriber.php
@@ -1,9 +1,9 @@
 <?php
 namespace WP_Rocket\Subscriber\Optimization;
 
+use WP_Rocket\deprecated\DeprecatedClassTrait;
 use WP_Rocket\Optimization\Remove_Query_String;
 use WP_Rocket\Subscriber\Optimization\Minify_Subscriber;
-use WP_Rocket\Traits\DeprecatedClass;
 
 /**
  * Hooks into WordPress to remove query strings for static files.
@@ -14,7 +14,7 @@ use WP_Rocket\Traits\DeprecatedClass;
  * @deprecated
  */
 class Remove_Query_String_Subscriber extends Minify_Subscriber {
-	use DeprecatedClass;
+	use DeprecatedClassTrait;
 
 	/**
 	 * Remove Query String instance.
